+++ scripts/macros/2014/monsterFeatures/generic/touchDamage.js
@@ -4,7 +4,8 @@ async function hit({trigger:{entity: item, token}, workflow}) {
     let config = itemUtils.getGenericFeatureConfig(item, 'touchDamage');
     let range = config.range;
     if (range > 0 && tokenUtils.getDistance(token, workflow.token) > range) return;
-    await workflowUtils.syntheticItemRoll(item, [workflow.token]);
+    if (item.system.uses.max && !item.system.uses.value) return;
+    await workflowUtils.syntheticItemRoll(item, [workflow.token], {consumeResources: true, consumeUsage: true});
 }
 export let touchDamage = {
     name: 'Touch Damage',